🫧 refactor: 네비게이션 네이밍 개선
